[EH] modify user register to auth route
