+++ README.md
@@ -1,6 +1,7 @@
 # AjouSelves_Backend
-* * *
+
 ## initial setting
+
 ```
 #config.js create
   server_port: 3000,
@@ -8,10 +9,14 @@
   db_host:'127.0.0.1',
   db_pw: input your password,
   db_user:input your user,
-  db_port:3306, 
+  db_port:3306,
   db_name: input your DB_NAME,
 ```
+
+---
+
 ## usage
+
 ```
 > 결과 확인은 postman 사용
 cd AjouSelves_Backend
